@@ -1,6 +1,6 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 	<PropertyGroup>
-		<TargetFrameworks>net6.0-windows10.0.19041; net8.0-windows10.0.19041; net9.0-windows10.0.19041;</TargetFrameworks>
+		<TargetFrameworks>net8.0-windows10.0.19041; net9.0-windows10.0.19041;</TargetFrameworks>
 		<PackageId>DSoft.System.Mvvm.UI.WinUI</PackageId>
 		<Version>2.2.2102.142-prerelease</Version>
 		<Description>WinUI 3 platform support for System.Mvvm.UI for NET6+ and UWP</Description>
@@ -38,13 +38,6 @@
 		<None Include="readme.md" Pack="true" PackagePath="\" />
 		<Compile Include="**\*.shared.cs" />
 		<Compile Include="**\*.shared.*.cs" />
-	</ItemGroup>
-
-	<ItemGroup Condition=" $(TargetFramework.StartsWith('net6.0-windows10')) ">
-		<PackageReference Include="Microsoft.WindowsAppSDK" Version="1.6.240923002" />
-		<PackageReference Include="Microsoft.Windows.SDK.BuildTools" Version="10.0.26100.1742" />
-		<Compile Include="**\*.uwp.cs" />
-		<Compile Include="**\*.uwp.*.cs" />
 	</ItemGroup>
 
 	<ItemGroup Condition=" $(TargetFramework.StartsWith('net8.0-windows10')) ">
