@@ -1,6 +1,6 @@
 ﻿<Project Sdk="MSBuild.Sdk.Extras/3.0.44">
 	<PropertyGroup>
-		<TargetFrameworks>net462;net6.0-windows7.0;net6.0-windows10.0.18362.0;net8.0-windows7.0;net8.0-windows10.0.18362.0;net9.0-windows7.0;net9.0-windows10.0.18362.0;</TargetFrameworks>
+		<TargetFrameworks>net462;net8.0-windows7.0;net8.0-windows10.0.18362.0;net9.0-windows7.0;net9.0-windows10.0.18362.0;</TargetFrameworks>
 		<GenerateAssemblyInfo>true</GenerateAssemblyInfo>
 		<PackageId>DSoft.System.Mvvm.UI.WPF</PackageId>
 		<Version>2.2.2102.142-prerelease</Version>
